# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Desafio backend capyba",
        "description": "Documentação da api desenvolvida para desafio backend para capyba - Gabriel Lima",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/vacancy": {
            "get": {
                "tags": [
                    "Vagas de emprego"
                ],
                "description": "Listagem de vagas de emprego",
                "parameters": [
                    {
                        "name": "hiring_mode",
                        "in": "query",
                        "description": "Tipo de contratação",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "default": "BOTH",
                                "enum": [
                                    "BOTH",
                                    "CLT",
                                    "PJ"
                                ]
                            }
                        }
                    },
                    {
                        "name": "is_home_office",
                        "in": "query",
                        "description": "Boolean representando se é home office ou não",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "salary",
                        "in": "query",
                        "description": "Tipo de ordenação a partir do valor do salario",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "default": "asc",
                                "enum": [
                                    "asc",
                                    "desc"
                                ]
                            }
                        }
                    },
                    {
                        "name": "created_at",
                        "in": "query",
                        "description": "Tipo de ordenação a partir da data de criação da vaga",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "default": "asc",
                                "enum": [
                                    "asc",
                                    "desc"
                                ]
                            }
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Pequisa textual para encontrar vaga",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "occupation",
                        "in": "query",
                        "description": "Tipo de vaga, back, front ou full stack",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "default": "FULL",
                                "enum": [
                                    "BACK",
                                    "FRONT",
                                    "FULL"
                                ]
                            }
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Pagina da pesquisa",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Quantidade de itens por pagina",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessed Entity"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/vacancy/my": {
            "get": {
                "tags": [
                    "Vagas de emprego"
                ],
                "description": "Listagem de vagas de emprego criadas pelo usuario",
                "operationId": "e3fbf5d82abafad979466c5bd3a56c97",
                "parameters": [
                    {
                        "name": "hiring_mode",
                        "in": "query",
                        "description": "Tipo de contratação",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "default": "BOTH",
                                "enum": [
                                    "BOTH",
                                    "CLT",
                                    "PJ"
                                ]
                            }
                        }
                    },
                    {
                        "name": "is_home_office",
                        "in": "query",
                        "description": "Boolean representando se é home office ou não",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "salary",
                        "in": "query",
                        "description": "Tipo de ordenação a partir do valor do salario",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "default": "asc",
                                "enum": [
                                    "asc",
                                    "desc"
                                ]
                            }
                        }
                    },
                    {
                        "name": "created_at",
                        "in": "query",
                        "description": "Tipo de ordenação a partir da data de criação da vaga",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "default": "asc",
                                "enum": [
                                    "asc",
                                    "desc"
                                ]
                            }
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Pequisa textual para encontrar vaga",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "occupation",
                        "in": "query",
                        "description": "Tipo de vaga, back, front ou full stack",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "default": "FULL",
                                "enum": [
                                    "BACK",
                                    "FRONT",
                                    "FULL"
                                ]
                            }
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Pagina da pesquisa",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Quantidade de itens por pagina",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessed Entity"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/vacancy/": {
            "post": {
                "tags": [
                    "Vagas de emprego"
                ],
                "description": "EP para alterar senha de um usuario",
                "operationId": "fb51db9ddaa6c29420af11bf463d8335",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "salary",
                                    "is_home_office",
                                    "occupation",
                                    "city_id",
                                    "hiring_mode"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "titulo da vaga",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Descrição da vaga",
                                        "type": "string"
                                    },
                                    "salary": {
                                        "description": "Valor da vaga",
                                        "type": "number"
                                    },
                                    "is_home_office": {
                                        "description": "Boolean reprsentando se a vaga é home office",
                                        "type": "boolean"
                                    },
                                    "occupation": {
                                        "description": "Ocupação [BACK, FRONT ou FULL]",
                                        "type": "boolean"
                                    },
                                    "city_id": {
                                        "description": "Id da cidade da empresa da vaga",
                                        "type": "integer"
                                    },
                                    "hiring_mode": {
                                        "description": "Tipo de contratação [CLT, PJ ou BOTH]",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessed Entity"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Para conseguir esse token, logue com email e senha com uma conta com email verificado",
                "name": "Bearer token",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "codeAuth": {
                "type": "http",
                "description": "Para conseguir esse token, crie uma conta, ou envie novamente um codigo para email de uma conta já cadastrada, que esse token virá na resposta da requisição",
                "name": "Bearer token",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}